# docs/api/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "网格社区治理平台 API",
    "description": "网格社区治理平台后端接口文档 v1.2.0 — 完整覆盖所有控制器",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/drone/stream/proxy/{deviceSn}/hls.m3u8": {
      "get": {
        "tags": [
          "无人机"
        ],
        "summary": "HLS视频流代理",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceSn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功（m3u8播放列表）"
          }
        }
      }
    },
    "/drone/stream/proxy/{deviceSn}/seg": {
      "get": {
        "tags": [
          "无人机"
        ],
        "summary": "TS分片代理",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceSn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "u",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功（视频流）"
          }
        }
      }
    },
    "/drone/stream/flv/{deviceSn}": {
      "get": {
        "tags": [
          "无人机"
        ],
        "summary": "FLV实时流地址",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceSn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/video/cameras/{id}/stream": {
      "get": {
        "tags": [
          "视频监控"
        ],
        "summary": "监控视频流地址",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}